FIX :: answerDelete와 commentDelete를 통합
