+++ src/components/layout/home/PetitionDetail/Comment/index.tsx
@@ -1,13 +1,10 @@
 import ProfileSvg from 'assets/profile.svg';
 import { FormatDatetime } from 'utils/FormatDatetime';
 import CheckSvg from 'assets/check.svg';
-import { useRecoilValue } from 'recoil';
-import { userState } from 'atoms/user.atom';
 import { CommentType } from 'types/petition.type';
 import Dialog from 'components/common/Dialog';
 import { useDialog } from 'hooks/useDialog';
-import { CommentDeleteFeature } from 'features/home/commentDelete.feature';
-import { AnswerDeleteFeature } from 'features/home/deleteAnswer.feature';
+import { ReplyDeleteFeature } from 'features/home/ReplyDelete.feature';
 import * as S from './style';
 
 const Comment = ({
@@ -19,8 +16,7 @@ const Comment = ({
 }: CommentType) => {
   const { date, time } = FormatDatetime(createdAt);
   const { closeDialog, openDialog } = useDialog();
-  const { deleteCommnetSubmit } = CommentDeleteFeature(id);
-  const { deleteAnswerSubmit } = AnswerDeleteFeature(id);
+  const { deleteSubmit } = ReplyDeleteFeature({ id, option });
 
   return (
     <>
@@ -48,33 +44,18 @@ const Comment = ({
           <S.Pre>{comment}</S.Pre>
         </S.Content>
       </S.Comment>
-      {option === 'STUDENT_COUNCIL' ? (
-        <Dialog
-          option="CONFIRM"
-          title="댓글 삭제"
-          content="정말 댓글을 삭제하시겠습니까?"
-          canceltext="취소"
-          checktext="삭제"
-          cancel={closeDialog}
-          check={() => {
-            deleteAnswerSubmit();
-            closeDialog();
-          }}
-        />
-      ) : (
-        <Dialog
-          option="CONFIRM"
-          title="댓글 삭제"
-          content="정말 댓글을 삭제하시겠습니까?"
-          canceltext="취소"
-          checktext="삭제"
-          cancel={closeDialog}
-          check={() => {
-            deleteCommnetSubmit();
-            closeDialog();
-          }}
-        />
-      )}
+      <Dialog
+        option="CONFIRM"
+        title="댓글 삭제"
+        content="정말 댓글을 삭제하시겠습니까?"
+        canceltext="취소"
+        checktext="삭제"
+        cancel={closeDialog}
+        check={() => {
+          deleteSubmit();
+          closeDialog();
+        }}
+      />
     </>
   );
 };
