--- src/features/home/commentDelete.feature.ts
@@ -1,24 +1,0 @@
-import { useMutation, useQueryClient } from 'react-query';
-import { deleteCommentPetition } from 'apis/petition.api';
-import { useSuccesToast } from 'hooks/useToast';
-import * as KEY from 'constants/key.constant';
-
-export const CommentDeleteFeature = (id: number) => {
-  const queryClient = useQueryClient();
-
-  const deleteMutate = useMutation(deleteCommentPetition, {
-    onSuccess: () => {
-      useSuccesToast('삭제 성공');
-      queryClient.invalidateQueries([KEY.PETITION]);
-    },
-    onError: (err) => {
-      console.log(err);
-    },
-  });
-
-  const deleteCommnetSubmit = () => {
-    deleteMutate.mutate(id);
-  };
-
-  return { deleteCommnetSubmit };
-};
